+++ api/internal/delivery/graph/dto.go
@@ -23,7 +23,7 @@ func buildSoundtrackModelV2(s *soundtrackEntities.Soundtrack) models.Soundtrack 
 		ID:        utils.IntToString(s.ID),
 		Title:     s.Title,
 		Author:    s.Author,
-		Duration:  s.Duration,
+		Duration:  int64(s.Duration),
 		CoverURL:  s.CoverImage,
 		AudioURL:  s.Audio,
 		Validated: false, // todo: remove
